requirements.txt
Flask==2.3.2
Pillow==9.5.0
rembg==2.0.50
Werkzeug==2.3.6
